split http and requests handling across multiple files
Now there is a dedicated file containing the relevant logic
for mobile metrics and for frontend data.
