+++ go/fhttpd/fhttpd.go
@@ -1,12 +1,7 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strconv"
@@ -71,168 +66,6 @@ func parseArgs() {
 
 var wg sync.WaitGroup
 
-type (
-	stringMap map[string]interface{}
-	stringSet map[string]bool
-)
-
-func (sm stringMap) DeleteString(k string) string {
-	v, _ := sm[k].(string)
-	delete(sm, k)
-	return v
-}
-
-// URL params which neeed to be converted to integer for JSON
-var integerKeyList = []string{"viewport_height", "viewport_width", "html_nodes", "script_nodes", "style_nodes", "v"}
-
-// Lookup table for those params
-var integerKeys stringSet
-
-func init() {
-	integerKeys = make(stringSet)
-	for _, k := range integerKeyList {
-		integerKeys[k] = true
-	}
-}
-
-func parseValue(k string, v string) (interface{}, error) {
-	if integerKeys[k] {
-		return strconv.Atoi(v)
-	}
-	return v, nil
-}
-
-func parseQuery(r *http.Request) (stringMap, error) {
-	sm := make(stringMap)
-	for k, v := range r.URL.Query() {
-		switch len(v) {
-		case 1:
-			v, err := parseValue(k, v[0])
-			if err != nil {
-				return sm, err
-			}
-			sm[k] = v
-		case 0:
-			sm[k] = ""
-		default:
-			return sm, fmt.Errorf("Parameter %s specified more than once", k)
-		}
-	}
-	return sm, nil
-}
-
-func extractFrontendData(r *http.Request) (stringMap, *util.RequestId, error) {
-	sm, err := parseQuery(r)
-	if err != nil {
-		return sm, nil, err
-	}
-	// add timestamps
-	now := time.Now()
-	sm["started_ms"] = now.UnixNano() / int64(time.Millisecond)
-	sm["started_at"] = now.Format(time.RFC3339)
-
-	// check protocol version
-	if sm["v"] == nil {
-		return sm, nil, errors.New("Missing protocol version number: v=1")
-	}
-	if sm["v"].(int) != 1 {
-		return sm, nil, fmt.Errorf("Unsupported protocol version: v=%d", sm["v"].(int))
-	}
-	// check logjam_action
-	if sm["logjam_action"] == nil {
-		return sm, nil, errors.New("Missing field: logjam_action")
-	}
-	// check request_id
-	if sm["logjam_request_id"] == nil {
-		return sm, nil, errors.New("Missing field: logjam_request_id")
-	}
-	id := sm["logjam_request_id"].(string)
-	// extract app and environment
-	rid, err := util.ParseRequestId(id)
-	if err != nil {
-		return sm, nil, err
-	}
-	sm["user_agent"] = r.Header.Get("User-Agent")
-	// log.Info("SM: %+v, RID: %+v", sm, rid)
-	return sm, rid, nil
-}
-
-func sendFrontendData(rid *util.RequestId, msgType string, sm stringMap) error {
-	appEnv := rid.AppEnv()
-	routingKey := rid.RoutingKey("frontend", msgType)
-	data, err := json.Marshal(sm)
-	if err != nil {
-		return err
-	}
-	publisher.Publish(appEnv, routingKey, data, util.NoCompression)
-	return nil
-}
-
-func writeErrorResponse(w http.ResponseWriter, txt string) {
-	statsMutex.Lock()
-	httpFailures++
-	statsMutex.Unlock()
-	http.Error(w, "400 RTFM", 400)
-	fmt.Fprintln(w, txt)
-}
-
-func writeImageResponse(w http.ResponseWriter) {
-	w.WriteHeader(200)
-	w.Header().Set("Cache-Control", "private")
-	w.Header().Set("Content-Type", "image/gif")
-	w.Header().Set("Content-Disposition", "inline")
-	w.Header().Set("Content-Transfer-Encoding", "base64")
-	io.WriteString(w, "R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==")
-}
-
-func serveFrontendRequest(w http.ResponseWriter, r *http.Request) {
-	defer recordRequestStats(r)
-	sm, rid, err := extractFrontendData(r)
-	if err != nil {
-		writeErrorResponse(w, err.Error())
-		return
-	}
-	msgType := extractFrontendMsgType(r)
-	err = sendFrontendData(rid, msgType, sm)
-	if err != nil {
-		writeErrorResponse(w, err.Error())
-		return
-	}
-	writeImageResponse(w)
-}
-
-func extractFrontendMsgType(r *http.Request) string {
-	if r.URL.Path == "/logjam/ajax" {
-		return "ajax"
-	}
-	if r.URL.Path == "/logjam/page" {
-		return "page"
-	}
-	return "unknown"
-}
-
-func serveAlive(w http.ResponseWriter, r *http.Request) {
-	defer recordRequestStats(r)
-	w.WriteHeader(200)
-	w.Header().Set("Cache-Control", "private")
-	w.Header().Set("Content-Type", "text/plain")
-	io.WriteString(w, "ALIVE\n")
-}
-
-func serveMobileMetrics(w http.ResponseWriter, r *http.Request) {
-	defer recordRequestStats(r)
-	bytes, err := ioutil.ReadAll(r.Body)
-	r.Body.Close()
-	if err != nil {
-		writeErrorResponse(w, err.Error())
-		return
-	}
-	appEnv := "mobile-production"
-	routingKey := "mobile"
-	publisher.Publish(appEnv, routingKey, bytes, util.NoCompression)
-	writeImageResponse(w)
-}
-
 func setupWebServer() *http.Server {
 	mux := http.NewServeMux()
 	mux.HandleFunc("/logjam/ajax", serveFrontendRequest)
@@ -243,36 +76,6 @@ func setupWebServer() *http.Server {
 	return &http.Server{
 		Addr:    spec,
 		Handler: mux,
-	}
-}
-
-func runWebServer(srv *http.Server) {
-	log.Info("starting http server on %s", srv.Addr)
-	if opts.KeyFile != "" && opts.CertFile != "" {
-		err := srv.ListenAndServeTLS(opts.CertFile, opts.KeyFile)
-		if err != nil && err != http.ErrServerClosed {
-			log.Error("Cannot listen and serve TLS: %s", err)
-		}
-	} else if opts.KeyFile != "" {
-		log.Error("cert-file given but no key-file!")
-	} else if opts.CertFile != "" {
-		log.Error("key-file given but no cert-file!")
-	} else {
-		err := srv.ListenAndServe()
-		if err != nil && err != http.ErrServerClosed {
-			log.Error("Cannot listen and serve: %s", err)
-		}
-	}
-}
-
-func shutdownWebServer(srv *http.Server, gracePeriod time.Duration) {
-	ctx, cancel := context.WithTimeout(context.Background(), gracePeriod)
-	defer cancel()
-	err := srv.Shutdown(ctx)
-	if err != nil {
-		log.Error("web server shutdown failed: %+v", err)
-	} else {
-		log.Info("web server shutdown successful")
 	}
 }
